Login and Logout

+++ index.js
@@ -1,7 +1,10 @@
 const express = require('express');
 const expressHandlebars = require('express-handlebars');
 const path = require('path');
-
+var url = require('url');
+const sessions = require("express-session");
+const cookieParser = require("cookie-parser");
+const uuidv4 = require('uuid').v4;
 const mysql = require('mysql2/promise');
 
 const PORT = process.env.PORT || 3001;
@@ -17,7 +20,15 @@ app.set('views', './views');
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(express.urlencoded({ extended: true }));
-
+app.use(express.json());
+app.use(cookieParser());
+app.use(sessions({
+    secret: "thisIsMySecretKey",
+    saveUninitialized: true,
+    resave: false,
+    name: 'Cookie de Sessao',
+    cookie: { maxAge: 1000 * 60 * 3 } // 3 minutos
+}));
 async function getConnection() {
     const connection = await mysql.createConnection({
         host: 'localhost',
@@ -36,8 +47,60 @@ async function query(sql = '', values = []) {
     return result[0];
 }
 
-app.get("/", async function (request, response) {
+app.use("*", async function (req, res, next) {
+    if (!req.session.usuario && req.cookies.token) {
+        const resultado = await query("SELECT * FROM usuarios WHERE token = ?", [req.cookies.token]);
+        if (resultado.length) {
+            req.session.usuario = resultado[0];
+        }
+    }
+    next();
+});
 
+app.get("/login", function (req, res) {
+    res.render("login", {
+        login: true,
+        tituloPagina: "Login",
+        titulo: "Login",
+        frase: "Utilize o formulário abaixo para realizar o login na aplicação."
+    });
+});
+
+app.post("/login", async function (req, res) {
+    const { user: usuario, pwd, keep_logged } = req.body;
+    const resultado = await query("SELECT * FROM usuarios WHERE email = ? AND senha = ?", [usuario, pwd]);
+    // console.log(resultado);
+
+    if (resultado.length > 0) {
+        if (keep_logged) {
+            const token = uuidv4();
+            console.log("----", token);
+            const isOk = await query("UPDATE usuarios SET token = ? WHERE usuario_id = ?", [token, resultado[0].usuario_id]);
+            console.log(isOk);
+            res.cookie("token", token);
+        }
+
+        req.session.usuario = resultado[0];
+        res.redirect("/");
+        return;
+    } else {
+        console.log("--666--");
+    }
+
+    res.render("login", {
+        login: true,
+        tituloPagina: "Login",
+        titulo: "Login",
+        frase: "Utilize o formulário abaixo para realizar o login na aplicação.",
+        mensagemErro: "Usuário/Senha inválidos!"
+    });
+});
+
+app.get("/", async function (request, response) {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
     const financas = await query('SELECT * FROM financas');
     const categorias = await query('SELECT * FROM categoria');
     let totalEntradas = 0;
@@ -52,13 +115,13 @@ app.get("/", async function (request, response) {
     const total = totalEntradas - totalSaidas;
     financas.map((financa) => {
         categorias.map((categoria) => {
-            if(financa.id_categoria === categoria.id_categoria) {
+            if (financa.id_categoria === categoria.id_categoria) {
                 financa.categoria = categoria.nome_categoria;
             }
         })
     })
     response.render('home', {
-        tituloPagina: 'Produtos',
+        tituloPagina: 'Finanças',
         listafinancas: financas,
         totalEntradas,
         totalSaidas,
@@ -67,6 +130,10 @@ app.get("/", async function (request, response) {
 });
 
 app.get("/excluir-produto/:id", async function (request, response) {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
     const id = parseInt(request.params.id);
     if (!isNaN(id) && id > 0) {
         await query('DELETE FROM financas WHERE id_financa = ?', [id]);
@@ -74,14 +141,11 @@ app.get("/excluir-produto/:id", async function (request, response) {
     response.redirect('/');
 });
 
-app.get('/buscar-produto/:nome', async (request, response) => {
-    let nome = request.params.nome;
-    let sql = "SELECT * FROM produtos WHERE nome_produto LIKE ?";
-    let valores = ['%' + nome + '%'];
-    const produtos = await query(sql, valores);
-});
-
 app.get('/editar-produto/:id', async function (request, response) {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
     const id = parseInt(request.params.id);
     const dadosProduto = await query("SELECT * FROM financas WHERE id_financa = ?", [id]);
     if (dadosProduto.length === 0) {
@@ -89,45 +153,85 @@ app.get('/editar-produto/:id', async function (request, response) {
         return;
     }
     const financa = dadosProduto[0];
-    console.log(financa);
+    const categorias = await query('SELECT * FROM categoria');
+    categorias.map((categoria) => {
+        if (categoria.id_categoria === financa.id_categoria) {
+            categoria.checked = true
+        }
+    })
     response.render('editar-produto', {
-        tituloPagina: 'Editar Produto',
-        financa
+        tituloPagina: 'Editar Finança',
+        financa,
+        categorias
     });
 });
 
 app.post('/editar-produto', async (request, response) => {
 
-    let { id, nome, preco, ativo } = request.body;
-    ativo = ativo ? 1 : 0;
-    const dadosPagina = {
-        tituloPagina: 'Editar Produto',
-        mensagem: '',
-        objProduto: { nome_produto: nome, id_produto: id, preco_produto: preco, ativo }
+    let id = request.body.id
+    let titulo = request.body.titulo;
+    let valor = request.body.valor;
+    let tipo = request.body.tipo;
+    let categoria = request.body.categoria;
+    let dataDate = new Date();
+    let dia = dataDate.getDay().toString();
+    let mes = dataDate.getMonth().toString();
+    let year = dataDate.getFullYear().toString();
+    let hora = dataDate.getHours().toString();
+    let minute = dataDate.getMinutes().toString();
+
+    if (dia.length == 1)
+        dia = '0' + dia;
+    if (mes.length == 1)
+        mes = '0' + mes;
+    if (hora.length == 1)
+        hora = '0' + hora;
+    if (minute.length == 1)
+        minute = '0' + minute;
+
+    let data = dia + "/" + mes + "/" + year;
+    let totalhora = hora + ":" + minute;
+
+    dadosPagina = {
+        tituloPagina: 'Editar Finança',
+        titulo,
+        valor,
+        categoria,
+        tipo,
+        hora,
+        totalhora
     }
 
     try {
-        if (!nome)
+        if (!titulo)
             throw new Error('Nome é obrigatório!');
-        if (preco <= 0)
-            throw new Error('Preço inválido!');
 
-        let sql = "UPDATE produtos SET nome_produto = ?, preco_produto = ?, ativo = ? WHERE id_produto = ?";
-        let valores = [nome, preco, ativo, id];
+        if (categoria === '0')
+            throw new Error('Categoria é obrigatório!');
+
+        if (valor <= 0)
+            throw new Error('Preco inválido!');
+
+        let sql = "UPDATE financas SET titulo = ?, valor = ?, tipo = ?, id_categoria = ?, data = ?, hora = ? WHERE id_financa = ?";
+        let valores = [titulo, valor, tipo === "entrada" ? 1 : 0, categoria, data, totalhora, id];
         await query(sql, valores);
-        dadosPagina.mensagem = 'Produto atualizado com sucesso!';
+        dadosPagina.mensagem = 'Produto Editado com sucesso!';
         dadosPagina.cor = "green";
     }
-    catch (e) {
-        dadosPagina.mensagem = e.message;
+    catch (error) {
+        dadosPagina.mensagem = error.message;
         dadosPagina.cor = "red";
     }
     response.render('editar-produto', dadosPagina);
-
 });
 
-app.get('/cadastrar-produto', function (request, response) {
-    response.render('cadastrar-produto', { tituloPagina: 'Cadastrar Produto' });
+app.get('/cadastrar-produto', async function (request, response) {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
+    const categorias = await query('SELECT * FROM categoria');
+    response.render('cadastrar-produto', { tituloPagina: 'Cadastrar Finança', categorias });
 });
 
 app.post('/cadastrar-produto', async (request, response) => {
@@ -156,7 +260,7 @@ app.post('/cadastrar-produto', async (request, response) => {
     let totalhora = hora + ":" + minute;
 
     dadosPagina = {
-        tituloPagina: 'Cadastrar Produto',
+        tituloPagina: 'Cadastrar Finança',
         titulo,
         valor,
         categoria,
@@ -164,10 +268,12 @@ app.post('/cadastrar-produto', async (request, response) => {
         hora,
         totalhora
     }
-
     try {
         if (!titulo)
             throw new Error('Nome é obrigatório!');
+
+        if (categoria === '0')
+            throw new Error('Categoria é obrigatório!');
 
         if (valor <= 0)
             throw new Error('Preco inválido!');
@@ -185,47 +291,88 @@ app.post('/cadastrar-produto', async (request, response) => {
     response.render('cadastrar-produto', dadosPagina);
 });
 
-app.get('/sobre', function (request, response) {
+app.get("/logout", function(req, res) {
+    res.cookie("token", "");
+    req.session.destroy();
+    res.redirect("/login");
+});
 
+app.get('/sobre', function (request, response) {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
     response.render('sobre', {
-        tituloPagina: 'Esta é a página Sobre',
-        nome: 'Ricardo',
-        idade: 35
+        tituloPagina: 'Sobre',
     });
 
 });
 
+app.get('/buscar-produto', async (request, response) => {
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
+    let nome = request.query.nome;
+    let sql = "SELECT * FROM financas WHERE titulo LIKE ?";
+    let valores = ['%' + nome + '%'];
+    const financas = await query(sql, valores);
+    const categorias = await query('SELECT * FROM categoria');
+    let totalEntradas = 0;
+    let totalSaidas = 0;
+    financas.map((fin) => {
+        if (fin.tipo == 1) {
+            totalEntradas += fin.valor;
+        } else {
+            totalSaidas += fin.valor;
+        }
+    })
+    const total = totalEntradas - totalSaidas;
+    financas.map((financa) => {
+        categorias.map((categoria) => {
+            if (financa.id_categoria === categoria.id_categoria) {
+                financa.categoria = categoria.nome_categoria;
+            }
+        })
+    })
+    response.render('home', {
+        tituloPagina: 'Finanças',
+        listafinancas: financas,
+        totalEntradas,
+        totalSaidas,
+        total,
+    })
+});
+
 app.get('/contato', function (request, response) {
-    response.render('contato');
+    if (!request.session.usuario) {
+        response.redirect("/login");
+        return;
+    }
+    response.render('contato', { tituloPagina: "Contato" });
 })
 
-app.post('/contato', function (request, response) {
+app.post('/contato', async function (request, response) {
 
-    let { nome, email, idade, linguagens } = request.body;
+    let { nome, email, mensagem } = request.body;
+    let mensagemErro = ""
 
-    let dadosRender = null;
+    if (nome.length < 3)
+        mensagemErro += 'Nome precisa ter pelo menos 3 letras!   -  ';
 
-    try {
+    if (mensagem.length < 20)
+        mensagemErro += 'Mensagem precisa ter pelo menos 20 letras!   -  ';
 
-        dadosRender = {
-            dadosValidos: true, nome, email, idade, linguagens
-        };
+    if (!email)
+        mensagemErro += 'E-mail é inválido !   -  ';
 
-        if (nome.length < 3) {
-            throw new Error('Nome precisa ter pelo menos 3 letras!');
-        }
-
-        if (!email) throw new Error('E-mail é inválido!');
-
+    if (mensagemErro === "") {
+        let sql = "INSERT INTO contato(nome, email, mensagem) VALUES(?, ?, ?)";
+        let valores = [nome, email, mensagem];
+        await query(sql, valores);
+        mensagemErro = "Contato salvo no Banco"
     }
-    catch (e) {
-        dadosRender = {
-            dadosValidos: false,
-            mensagemErro: e.message
-        }
-    }
-
-    response.render('contato', dadosRender);
+    response.render('contato', { tituloPagina: "Contato", mensagemErro });
 
 })
